requirements.txt
python-rtmidi
pywin32
playsound
comtypes
pycaw
phue
infi.systray